bug fixing and renaming json to json_utils as it collides with json base package
