+++ nexus_constructor/component_type.py
@@ -17,8 +17,10 @@ COMPONENT_TYPES = {
     "NXcapillary",
     "NXcollimator",
     "NXcrystal",
+    "NXdata",
     "NXdetector_group",
     "NXdetector_module",
+    "NXentry",
     "NXfermi_chopper",
     "NXfilter",
     "NXflipper",
